Account managment done and working
cards managment, menus, buttons

+++ src/controller/MinBalanceChangeController.java
@@ -17,7 +17,7 @@ public class MinBalanceChangeController extends DataChangeController {
 	@Override
 	void changeData(String newData) throws CustomException {
 		try {
-			double minBalance = Double.parseDouble(newData);
+			int minBalance = Integer.parseInt(newData);
 			if (minBalance < account.getBalance()) {
 				account.setMinBalance(minBalance);;
 			} else {
